update wheels and tests

+++ scripts/update_wheels.py
@@ -1,12 +1,12 @@
 import os
-import posixpath
 import shutil
+from contextlib import suppress
 from pathlib import Path
 from typing import Callable, Optional, Tuple
 
 import toml
 from config import Config
-from jampy_utils.path import parse_path_str
+from jampy_util.path import normalize_path
 
 
 def parse_wheel_path(
@@ -27,33 +27,45 @@ def parse_wheel_path(
 
 def update_wheels(package_name, src_file_path: Path) -> None:
     dst_dir = Config.PROJECT_ROOT.joinpath('wheels')
-    for wheel_path in dst_dir.glob(f"{package_name}-*.whl"):
-        os.remove(wheel_path)
-    shutil.copy2(src_file_path, dst_dir)
+    if (dst_dir.exists() is True) and (dst_dir.is_dir() is True):
+        for wheel_path in dst_dir.glob(f"{package_name}-*.whl"):
+            os.remove(wheel_path)
+        shutil.copyfile(src_file_path, dst_dir.joinpath(src_file_path.name))
+        return
+    with suppress(FileNotFoundError):
+        os.remove(dst_dir)
+    dst_dir.mkdir()
+    shutil.copyfile(src_file_path, dst_dir.joinpath(src_file_path.name))
 
 
 def update_install_wheels_bash(package_name: str) -> None:
     bash_path = Config.PROJECT_ROOT.joinpath('install_wheels.sh')
+    bash_lines = [
+        "#!/bin/bash",
+        "python scripts/update_wheels.py",
+    ]
     commands = [
         x.strip()
         for x in bash_path.read_text().splitlines()
         if (x.startswith('pip') and package_name not in x)
     ]
     commands.append(f'pip install --find-links="./wheels" {package_name}')
-    bash_path.write_text("#!/bin/bash\n" + "\n".join(commands))
+    bash_lines.extend(commands)
+    bash_path.write_text("\n".join(bash_lines))
 
 
-def main(repo_path_str: str) -> None:
-    _, repo_dir = parse_path_str(repo_path_str)
+def main(package_path_str: str) -> None:
+    _, repo_dir = normalize_path(package_path_str)
     package_name, _, wheel_path = parse_wheel_path(repo_dir)
     if wheel_path is None or package_name is None:
-        raise FileNotFoundError(f'wheels of {repo_path_str} not found!')
+        raise FileNotFoundError(f'wheels of {package_path_str} not found!')
     update_wheels(package_name, wheel_path)
     update_install_wheels_bash(package_name)
 
 
 if __name__ == '__main__':
-    for repo in [
-        "~/repos/jampy_utils",
+    for package in [
+        "~/repos/sheldon/jampy_util/jampy_util",
+        "~/repos/sheldon/jampy_util/jampy_util_intelligence",
     ]:
-        main(repo)
+        main(package)
